Backend/requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
defusedxml==0.7.1
distlib==0.3.7
Django==5.0.4
django-allauth==0.60.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.49.0
idna==3.6
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mongoengine==0.28.2
mysqlclient==2.2.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==3.11.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.12.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tk==0.1.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
Werkzeug==3.0.2
